Subblock-subblock, stat report
subblock-subblock logic implemnted completely, looks great and scaleable, but its the end of the day and testing is giving weird results hence will look at it tomorrow. Status report was completed yesterday, yesterday also i worked on the code to get the coordinated of the created block for sub-sub comparison. Today i theoretically finished the second part of it. Also helped steve with some csv stuff.