EOD
In the morning tried to optimise code, I have a commit for that.

Then mainly time was spent finding this function in trimesh library and comparing its use in different versions.

Then made an exmaple to verify if trimesh is working correctly. eg was: Used the points and sold martin had given me before and then used trimesh to find the points within the solid and then made a new pointset using sdk and then viewed it on the viewer to verify.

Then started thinking about the ways I can use 3d solid comparison. By that I meant:
Do we want to compare if the domain of these points match to what the subblocked model given by DomainMCF is or do we wanna compare the blocks these points are in for two subblock models having this solid and made of these points?

Then martin suggested the first approach should be:
1.Generate grid of smallest sub-block between models
2.Filter those centroids to only include those within the solid
3.Proceed like normal to compare domains for each of those points

So generated a couple new models and started doing that, then I realised I also had to convert the coordinates of the solid to normal coordinates from world coordinates before doing the contains_points check, it is now 6:22, so ending my day for now.