EOD
Done with sub-sub 3-d, just gotta do limiting solids testing and then complete the deletion part.

spend the first half of the day looking at how to improve efficiency after seeing andrew's post on jira. I have already implemented a np.where()[0] which is better than nonzero. also np.all seems to be the best way to implement stuff as well. [Formulate this better]

Also looked into the filtering part in sub-sub and ran across the following problem:
 if I wanna add the filter feature to sub-sub. I have to translate the visibility of the blocks on the viewer to my updated smaller blocks.
Now to do that, the way I can think of to proceed is, I can make an updated domain array in which the blocks whose visibility is affected will be categorised as "filtered" instead of their respective domain and everything else is done the same way.
The only thing is, this doesn't contribute to increasing efficiency at all as we will be parsing through the same number of centroids. can you think of a way through which I can translate the visibility of the block model to the updated smaller blocks?

Need to come up with something here as well.