EOD
First half of the day was spent optimising the code, create a dict of reverse grid_index and got it working, it increased the efficiency quite a bit, then showed martin the code and how its increasing the efficiency.

However the reverse-grid_index is helpful because in sub-sub comparison we are creating smaller number of blocks and hence increasing the total number of blocks but we are not using the grid index function to find which block they are located in.

However we have to use the grid_index function to make the reverse grid_index and iterate through all parent blocks for that.

So to optimise the optimised code further, it is necessary to get rid of the grid_index function to create the reverse_grid_index. I had worked on that before, but was not able to find a proper method.

Then I showed martin some of the old methods that I had tried and used but failed, so we looked into one of those methods and tried to implement it, and understand where it was failing but could not get it working.

Then Martin suggested to do some more testing on whether the optimised code works everywhere.

While doing that I found a error it throws in certain cases, I came up with a couple of approaches to solve it and I am working on which approach will keep the code most efficient.

Tomorrow, I am hoping to have a talk with Richard and Andrew to decide what to work on next.