EOD - exe
So today I spent time fighting the evils of rtree in the trimesh library
which is being used for the 3d solid filter, it basically worked
differntly for the nb and the normal py file.Then I was
extinguishing any fires caused by pyinstaller like first it not
installing mapteksdk, then it giving a huge onefile which would take
forever to unpack.

Then after everything was sorted I sent it to Martin so that he could
run it on his PC, I found multiple internal problems which are easily
fixable. However we still think the code runs slowly on big models. The
good part is this method works and by just clciking on the file it
generates a graph for you on spot, so it will quite useful as well as
easy to use.