EOD
Made quite some changes, was mainly working on the first talk I talked baout today in the morning which is basically figuring out if offset is needed and how to implement it (give proper explanation), I generated some models to do some testing, but could not get through it completely because of memoryerrors. Hopefully I ll get this done today.

Also scheduled my presentation to be on 15th Feb.