EOD
Spent a lot of time on optimisation, found this numba thing, still need to dig into this further to make it work.

Then got onto the assumptions thing again, did some testing and have to make some chnages to the code so still working on that.

Alos found out about the correlation of the subblock cooridnates and the number of new blocks created, need to have proper think about how to control that.